+++ internal/lpcnetplc/state.go
@@ -645,13 +645,13 @@ func quantizePCMInt16LikeInPlace(frame []float32) {
 }
 
 func quantizePCMInt16Like(sample float32) float32 {
-	v := float64(sample) * 32768
+	v := sample * 32768
 	if v < -32767 {
 		v = -32767
 	}
 	if v > 32767 {
 		v = 32767
 	}
-	v = math.Floor(0.5 + v)
-	return float32(v * (1.0 / 32768.0))
+	v = float32(math.Floor(0.5 + float64(v)))
+	return v * (1.0 / 32768.0)
 }
